Refine background tests (compile, online update)
Implement separate class BackgroundJobs for background jobs like these.
Notes for changes in online update :
Existing solution was using signal, but it was kind of limited and even
unsafe because strictly speaking we are only allowed to use async-signal
safe functions inside signal handler. So instead in this new
implementation named pipe (a.k.a FIFO) was used for notifying online
updater monitor of the file user wants to be used for updating.
Background compile test is essentially unchanged from the original
implementation. (Only, codes were moved into a dedicated class)