Apply rematch2 API changes, Fix bug regarding test result update.
The bug was that test result was not properly updated in such a case
where updated result differs from the previous ones.
This was caused by misunderstanding of how pointer dereference operator
of select cursor works. Actually each dereference is different instance
which is why previous implementation did not update the TestResult table
when it should do.