Revamp of PcreChecker part 1
Now db_setup.* are removed and the contents are essentially merged to
PcreChecker class which is now in PcreChecker.*.
main.cpp is added to explicitly process command line options.

Strengthen the command line capabilities.
Capability of writing contents of DB tables to an output json is added.
Tried to remove redundant codes as much as possible using templates.

TBD
- write tests table related items to json file
- for now CheckerShell capability is temporarily disabled, need to
  enable it back